优化数组存储 1

+++ runtime/Player_Partial_RunMethod.cs
@@ -366,7 +366,7 @@ namespace juicescript.runtime
 				m_scopePayload.methodFlags = method.Flags;
 
 				m_scopePayload.mScopePtr = (byte)mScopeId;
-				m_scopePayload.__sendargcount = args;
+				m_scopePayload.__sendargcount = (byte)args;
 
 				//save this 
 				{
@@ -1094,7 +1094,7 @@ namespace juicescript.runtime
 			m_scopePayload.methodFlags = method.Flags;
 
 			m_scopePayload.mScopePtr = (byte)mScopeId;
-			m_scopePayload.__sendargcount = argsCount;
+			m_scopePayload.__sendargcount = (byte)argsCount;
 
 
 			//save this
